warlock: fix pet base stats, including crit
For melee crit, the soul corruption debuff (spellId: 25805) was used to
reduce the agility of the pets to 0, leaving the result of
`GetCritChanceFromAgility("pet")` as the base crit value.

For spell crit we just calculated the base value assuming every point of
intellect gives the same amount of crit (so full benefit from the first
20 points that give reduced mana) using `GetSpellCritChanceFromIntellect("pet")`
and 166.6666 intellect required for 1% crit.

Pet base damage was calculated indirectly, essentially being the
leftover after all the damage increase from AP is taken into account.
Note that at this time on the beta pets AP scaling is weird, they seem
to have hidden modifiers on their damage that should not be there. These
modifiers were accounted for, such that AP scaling is the expected 14 AP
/ DPS, and we assume the base damage was affected by this too which may
prove incorrect. See JamminL/wotlk-classic-bugs#328