warlock: add pet expertise scaling
It's currently not well understood what the exact formula is for warlock
pets expertise scaling. What we do know is that suppression is not
considered and that it caps out at 17% (= -6.5% dodge).

Also there is good evidence that the scaling is continous, and not in
discrete 0.25% steps. Currently expertise is rounded down making it
discrete so this needs to be fixed later.