warlock: fix pets being intermixed on 'Automatic'
When automatic mode was selected, and a pet that is not the typical pet
for that spec what would happen is some parts of the code would swap
that pet out for the 'typical' choice while others would not, leading to
weird situations like imp meeleing.

We remove that code entirely since it serves no purpose anymore, should
also remove the UI options at some point since this was it's last use.