ui: change demonic pact raid buff value meaning
Previously, you'd set the amount equal to 10 times what you wanted the
demonic pact to be, with the reasoning that DP is 10% of the respective
warlocks spellpower. However this is problmatic for several reasons:
- DP is actually more than 10% on first application (11%)
- The ingame buff shows the SP amount
- For the demo sim we show average DP again as the actual SP amount

By limiting this value to 1000, it should be self-evident for users how
it works now, since most will want to sim with DP values of at least
200, which would translate to a value of 2000 in the old way, which can
no longer be set.

Fixes #2807.