Configure Blood DK Normal and Threat Openers to conditionally skip using Blood Tap and the subsequent Icy Touch enabled by it.
If the DK has the T10 4pc bonus enabled and does not schedule the first Blood Tap usage outside the time the opener uses Blood Tap + the Blood Tap Cooldown, the noted spells will not be used in either opener scenario.

This change comes with a caveat of potentially delaying scheduled Blood Tap casts if the opener Blood Tap is delayed due to ability misses.